Change type parameter for discharge setter

+++ island/src/main/java/island/Tile/River.java
@@ -25,7 +25,7 @@ public class River {
         return this.discharge;
     }
 
-    public void setDischarge(int val){
+    public void setDischarge(double val){
         this.discharge = val;
         this.humidity = this.discharge * 10;
     }
